pyproject.toml
[project]
name = "lincapr-experiments"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "biopython>=1.86",
    "dataclasses>=0.8",
    "matplotlib>=3.10.7",
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "pathlib>=1.0.1",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "tqdm>=4.67.1",
    "typing>=3.10.0.0",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
]
